changed parseInt to parseDouble
I forgot we were working with doubles. I'm dumb.

+++ com/meteorcode/JRPNMachine/RPNMachine.java
@@ -32,7 +32,7 @@ public class RPNMachine {
 				stack.push(new StackMember (Operation.MODULO));
 				break;
 			default:
-				stack.push(new StackMember (Integer.parseInt(inputString)));
+				stack.push(new StackMember (Double.parseDouble(inputString)));
 				break;
 		}
 	}
